♻️ Refactor modules used in rag server
- Unify rag and database folder
- Move crud functions to db_manager and retrieval
