--- database/__init__.py
@@ -1,3 +1,0 @@
-from .faiss_client import FAISSClient
-from .schemas import *
-from .crud import *
