+++ prompts/prompt.py
@@ -1,24 +1,8 @@
 from langchain_core.prompts.few_shot import FewShotPromptTemplate
 from langchain_core.prompts.prompt import PromptTemplate
 
-def load_template(template_name):
+def load_template(template_name, input_text):
     template_path = f"prompts/{template_name}.md"
     with open(template_path, "r", encoding="utf-8") as file:
         template = file.read()
-    return template
-
-def one(input_text):
-    prompt = load_template("test1")
-    return prompt.format(input=input_text)
-
-def two(input_text):
-    prompt = load_template("test2")
-    return prompt.format(input=input_text)
-
-def three(input_text):
-    prompt = load_template("test3")
-    return prompt.format(input=input_text)
-
-def extract_rss_content_prompt(input_text):
-    prompt = load_template("extract_rss_content_prompt")
-    return prompt.format(content=input_text)
+    return template.format(input=input_text)
