+++ utils/__init__.py
@@ -1,2 +1,7 @@
-from .processing import *
 from .llm import *
+from .embedding import *
+from .calculate_token import *
+from .processing import *
+
+from dotenv import load_dotenv
+load_dotenv()
